Rename utils.ts to common.ts
