+++ widgets/nodes/src/common.ts
